marketplace/pyproject.toml
[tool.poetry]
name = "fair_python_market-place"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
django-mptt = "^0.13.4"
pika = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
